OCR and Text Summarization/requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///C:/ci/aiohttp_1646806572557/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
atomicwrites==1.4.0
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
blinker==1.4
blis==0.7.7
Bottleneck @ file:///C:/ci/bottleneck_1648010904582/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
catalogue==2.0.7
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.12.0
conda-package-handling @ file:///C:/ci/conda-package-handling_1649106011304/work
cryptography @ file:///C:/ci/cryptography_1633520531101/work
cymem==2.0.6
DAWG-Python==0.7.2
debugpy @ file:///D:/bld/debugpy_1636043436560/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
frozenlist @ file:///C:/ci/frozenlist_1637767271796/work
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib==0.4.1
grpcio @ file:///C:/ci/grpcio_1637590978642/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.15.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562621412/work
iniconfig==1.1.1
ipykernel @ file:///D:/bld/ipykernel_1641325648894/work/dist/ipykernel-6.6.1-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1641404879599/work
jedi @ file:///D:/bld/jedi_1637175252639/work
Jinja2==3.1.2
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///D:/bld/jupyter_core_1636814453100/work
kaleido @ file:///home/conda/feedstock_root/build_artifacts/python-kaleido_1615204619408/work
keras==2.8.0
langcodes==3.3.0
langdetect==1.0.7
Markdown @ file:///C:/ci/markdown_1614364082838/work
MarkupSafe==2.1.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
menuinst==1.4.16
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict @ file:///C:/ci/multidict_1607349747897/work
murmurhash==1.0.7
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
networkx==2.8.4
nltk==3.7
numexpr @ file:///C:/ci/numexpr_1640704337920/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1649782970434/work
oauthlib @ file:///opt/conda/conda-bld/oauthlib_1644332107998/work
opencv-python==4.5.5.62
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1641443174429/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathy==0.6.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.0.1
plac==1.1.3
pluggy==1.0.0
preshed==3.0.6
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
protobuf==3.19.1
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.3.0/pt_core_news_sm-3.3.0-py3-none-any.whl
py==1.11.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.8.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
PyJWT @ file:///C:/ci/pyjwt_1619682721924/work
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pytesseract==0.3.9
pytest==7.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pyvirtualcam==0.9.0
pywin32==303
pyzmq @ file:///D:/bld/pyzmq_1635877415276/work
regex==2022.6.2
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.3.0/ru_core_news_sm-3.3.0-py3-none-any.whl
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
scipy @ file:///C:/bld/scipy_1644357595180/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==5.2.1
spacy==3.3.1
spacy-langdetect==0.1.2
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
tensorboard @ file:///C:/Users/san10428/Anaconda3/conda-bld/tensorboard_1640281199884/work/tensorboard-2.7.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/ci/tensorboard-data-server_1633035225228/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
thinc==8.0.17
tomli==2.0.1
torch==1.7.1
tornado @ file:///D:/bld/tornado_1635819745510/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typer==0.4.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wasabi==0.9.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.3.0/xx_ent_wiki_sm-3.3.0-py3-none-any.whl
xx-sent-ud-sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.3.0/xx_sent_ud_sm-3.3.0-py3-none-any.whl
yarl @ file:///C:/ci/yarl_1606940155993/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
